# Change reorder_dimensions behavior to favor output writting sequence